Make atom watchers atomic with respect to reset!.
In Atom.reset(), the old value was being stored, then the new value
set (non-atomically), allowing for the atom to change state between get
and set. The watchers were then notified with this potentially out of
date old value and the new value. This is in contrast to the behavior of
swap! with a (constantly newvalue), which did a compare and
set. Further, the validation was happening between get and set,
increasing the opportunity for thread interleaving.

This change makes Atom.reset() do a compareAndSet loop to ensure that
the old value is not out of date. Timing tests indicate that there is
very little to no time cost associated with this change.

There is also a test for the new atomic watch behavior.